refactor: remove message conversion validation and streamline connection handling in graph endpoints
