+++ core/src/ten_manager/src/designer/graphs/connections/msg_conversion/mod.rs
@@ -5,4 +5,3 @@
 // Refer to the "LICENSE" file in the root directory for more information.
 //
 pub mod update;
-pub mod validate;
