+++ core/src/ten_manager/src/designer/graphs/nodes/add.rs
@@ -181,7 +181,7 @@ pub async fn add_graph_node_endpoint(
         &request_payload.app,
         &request_payload.addon,
         &uri_to_pkg_info,
-        graph_info.app_base_dir.as_ref(),
+        &graph_info.app_base_dir,
         pkgs_cache,
     ) {
         // Validate the request payload before proceeding.
