+++ core/src/ten_manager/tests/test_case/designer/graphs/connections/msg_conversion/modify_with_msg_conversion.rs
@@ -1234,7 +1234,7 @@ mod tests {
         let body_str = std::str::from_utf8(&body).unwrap();
         println!("Response body: {}", body_str);
 
-        assert_eq!(body_str, "Failed to check message conversion schema: { .aaa: type is incompatible, source is [uint64], but target is [string] }");
+        assert_eq!(body_str, "Failed to validate connection schema: { .aaa: type is incompatible, source is [uint64], but target is [string] }");
     }
 
     #[actix_web::test]
